Added CheerLights mode
CheerLights (https://cheerlights.com/) is a community of connected lights around the world. The color is set by anyone on the Discord server and other social networks. The color is then broadcast to all the connected lights. The XMAS tree can connect to the CheerLights project when you select mode 5.